Moved fetchResourceData in public/utils
